Compressed almost all the scripts into one file, all test still passing, and 100% coverage
